comp

+++ src/main/java/com/project/work_hive/payloads/CompanyAdminId.java
@@ -13,6 +13,6 @@ import lombok.Setter;
 @AllArgsConstructor
 @EqualsAndHashCode
 public class CompanyAdminId implements Serializable {
-    private Long companyId;
-    private Long employeeId;
+    private int companyId;
+    private int employeeId;
 }
